Docs/crypto currency route

+++ src/crypto-currencies/crypto-currencies.service.ts
@@ -11,6 +11,7 @@ import { CryptoCurrency } from './entities/crypto-currency.entity';
 
 @Injectable()
 export class CryptoCurrenciesService {
+  // inject `CryptoCurrency` Repo to comunicate between server and database
   constructor(
     @InjectRepository(CryptoCurrency)
     private cryptoRepo: Repository<CryptoCurrency>,
@@ -20,21 +21,36 @@ export class CryptoCurrenciesService {
    Action:
       - insert new currency
       - return inserted currency
+    Exception:
+      - Currency not found (base on `db.name` column):
+          - throw `conflictException - 409`
   */
   async create(createCryptoCurrencyDto: CreateCryptoCurrencyDto) {
-    const userExists = await this.cryptoRepo.findOne({
+    // Step01 - try to get currency by `name` to check if exists
+    const currencyExists = await this.cryptoRepo.findOne({
       where: {
         name: createCryptoCurrencyDto.name,
       },
     });
 
-    if (userExists) {
+    if (currencyExists) {
+      // Step02_1 - if currency exists, return `409 - conflict` response to client
       throw new ConflictException();
     } else {
+      /* 
+      Step02_2 - if currency does not exists:
+        - create currency
+      */
       const newCrypto = this.cryptoRepo.create({
         ...createCryptoCurrencyDto,
         created_at: new Date(),
       });
+
+      /* 
+      Step3
+        - save `currency` to `db.currencies`
+        - return `currency` with `200 status code`
+      */
       return this.cryptoRepo.save(newCrypto);
     }
   }
@@ -57,9 +73,23 @@ export class CryptoCurrenciesService {
    Action:
       - update currency
       - return updated curreny
+    Exceptions:
+      - if currency not found:
+        - throw `NotFoundException - 404`
   */
   async update(id: number, updateCryptoCurrencyDto: UpdateCryptoCurrencyDto) {
+    // Step01 -> try to find currency base on id
     const user = await this.findOne(id);
+
+    /*
+      Step02:
+        - if currency exists:
+          - Update currency
+          - Save to DB
+          - return Updated currency to user
+        - if currency NOT exists:
+          - throw `NotFoundException - 404`
+    */
     return user
       ? this.cryptoRepo.save({ ...user, ...updateCryptoCurrencyDto })
       : this.CustomNotFoundException();
@@ -68,19 +98,31 @@ export class CryptoCurrenciesService {
   /* 
    Action:
       - remove
-      - return removed curreny
+      - return removed curreny to client
     Exceptions:
       - :id not found -> 404 not found error
   */
   async remove(id: number) {
-    const user = await this.findOne(id);
-    const removeResult = await this.cryptoRepo.delete(id);
+    // Step01 -> try to find currency base on id
+    const currency = await this.findOne(id);
 
-    return removeResult.affected != 0
-      ? { ...user, deleted_at: new Date() }
-      : this.CustomNotFoundException();
+    /*
+      Step02:
+        - if currency exists in DB:
+          - hard delete currency from DB
+          - return deleted currency to user with `deleted_at: Date`
+        - if currency NOT exists in DB:
+          - throw `NotFoundException - 404`
+    */
+    if (currency) {
+      await this.cryptoRepo.delete(id);
+      return { ...currency, deleted_at: new Date() };
+    } else {
+      throw new NotFoundException();
+    }
   }
 
+  // Action: throw `NotFoundException` for inline rules
   CustomNotFoundException() {
     throw new NotFoundException();
   }
